Features/Legal Updates/requirements.txt
# Python
python == 3.10

# Core
pymongo>=4.6.0
python-dotenv>=1.0.1

# Parsing & scraping
feedparser>=6.0.11
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Utilities
tqdm>=4.66.5